+++ app/Providers/AppServiceProvider.php
@@ -6,10 +6,13 @@ namespace App\Providers;
 
 use Domain\Admin\Models\Admin;
 use Domain\Blueprint\Models\Blueprint;
+use Domain\Page\Models\Page;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\Relation;
 use Illuminate\Support\ServiceProvider;
+use Illuminate\Support\Str;
 use Illuminate\Validation\Rules\Password;
+use Stancl\Tenancy\Database\Models\Tenant;
 
 /** @property \Illuminate\Foundation\Application $app */
 class AppServiceProvider extends ServiceProvider
@@ -20,14 +23,20 @@ class AppServiceProvider extends ServiceProvider
 
     public function boot(): void
     {
-        Model::preventLazyLoading( ! $this->app->isProduction());
-        Model::preventSilentlyDiscardingAttributes( ! $this->app->isProduction());
+        Model::shouldBeStrict( ! $this->app->isProduction());
+
+        Model::handleMissingAttributeViolationUsing(function (Model $model, string $key) {
+            if ($model instanceof Tenant && Str::startsWith($key, Tenant::internalPrefix())) {
+                return null;
+            }
+        });
 
         Relation::enforceMorphMap([
             Admin::class,
             config('permission.models.role'),
             config('tenancy.tenant_model'),
             Blueprint::class,
+            Page::class,
         ]);
 
         Password::defaults(