+++ database/migrations/tenant/2022_11_16_050908_create_pages_table.php
@@ -1,0 +1,31 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+use Domain\Blueprint\Models\Blueprint as BlueprintModel;
+
+return new class () extends Migration {
+    public function up(): void
+    {
+        Schema::create('pages', function (Blueprint $table) {
+            $table->id();
+
+            $table->foreignIdFor(BlueprintModel::class)->constrained();
+
+            $table->string('name')->unique();
+            $table->string('slug')->unique();
+
+            $table->json('data')->nullable();
+
+            $table->timestamps();
+        });
+    }
+
+    public function down(): void
+    {
+        Schema::dropIfExists('pages');
+    }
+};