+++ domain/Admin/Database/Factories/AdminFactory.php
@@ -36,6 +36,7 @@ class AdminFactory extends Factory
             'email_verified_at' => now(),
             'password' => 'secret',
             'remember_token' => Str::random(10),
+            'timezone' => config('domain.admin.default_timezone'),
         ];
     }