+++ tests/Feature/Filament/Resources/TenantResource/Pages/EditTenantTest.php
@@ -11,7 +11,7 @@ use function Pest\Laravel\assertDatabaseCount;
 use function Pest\Laravel\assertDatabaseHas;
 use function Pest\Livewire\livewire;
 
-beforeEach(fn () => loginAsAdmin());
+beforeEach(fn () => loginAsSuperAdmin());
 
 it('can render page', function () {
     $tenant = TenantFactory::new()->createOne();